[#279] refactor: import 문 변경 및 주석 제거
