+++ components/main/index.ts
@@ -1,3 +1,4 @@
 export { default as GoogleLoginButton } from "./googleLoginButton";
 export { default as KakaoLoginButton } from "./kakaoLoginButton";
 export { default as NaverLoginButton } from "./naverLoginButton";
+export { default as LoginButton } from "./loginButton";
